Fix tests and deprecations
