+++ .gitignore
@@ -3,3 +3,4 @@ composer.phar
 phpunit.xml
 vendor/
 .php_cs.cache
+.phpunit.result.cache
