+++ .travis.yml
@@ -14,28 +14,31 @@ env:
 matrix:
     fast_finish: true
     include:
-          # Minimum supported dependencies with the latest and oldest PHP version
-        - php: 7.2
-          env: COMPOSER_FLAGS="--prefer-stable --prefer-lowest"
+          # Minimum supported dependencies with the oldest PHP version
         - php: 7.1
-          env: COMPOSER_FLAGS="--prefer-stable --prefer-lowest"
+          env: COMPOSER_FLAGS="--prefer-stable --prefer-lowest" SYMFONY_PHPUNIT_VERSION=7.5
 
           # Test the latest stable release
         - php: 7.1
         - php: 7.2
         - php: 7.3
+        - php: 7.4
           env: COVERAGE=true PHPUNIT_FLAGS="-v --testsuite main --coverage-text --coverage-clover=build/coverage.xml"
 
         - php: 7.1
           # We need `--prefer-source`, otherwise the `Doctrine\Tests\OrmTestCase` class won't be found.
-          env: PHPUNIT_FLAGS="-v --testsuite doctrine" COMPOSER_FLAGS="--prefer-source" DEPENDENCIES="dunglas/symfony-lock:^3"
-        - php: 7.2
-          env: DEPENDENCIES="dunglas/symfony-lock:^3"
-        - php: 7.2
-          env: DEPENDENCIES="dunglas/symfony-lock:^4"
+          env: PHPUNIT_FLAGS="-v --testsuite doctrine" COMPOSER_FLAGS="--prefer-source" SYMFONY_VERSION=3.4.* SYMFONY_PHPUNIT_VERSION=7.5
+        - php: 7.4
+          env: SYMFONY_VERSION=3.4.*
+        - php: 7.4
+          env: SYMFONY_VERSION=4.3.*
+        - php: 7.4
+          env: SYMFONY_VERSION=4.4.*
+        - php: 7.4
+          env: SYMFONY_VERSION=5.0.*
 
           # Latest commit to master
-        - php: 7.2
+        - php: 7.4
           env: STABILITY="dev"
 
     allow_failures:
@@ -45,7 +48,7 @@ matrix:
 before_install:
     - if [[ $COVERAGE != true ]]; then phpenv config-rm xdebug.ini || true; fi
     - if ! [ -z "$STABILITY" ]; then composer config minimum-stability ${STABILITY}; fi;
-    - if ! [ -v "$DEPENDENCIES" ]; then composer require --no-update ${DEPENDENCIES}; fi;
+    - if [ "$SYMFONY_VERSION" != "" ]; then composer require "symfony/symfony:${SYMFONY_VERSION}" --no-update; fi;
 
 install:
     - composer update ${COMPOSER_FLAGS} --prefer-dist --no-interaction
