+++ Tests/DependencyInjection/ConfigurationTest.php
@@ -34,7 +34,7 @@ class ConfigurationTest extends TestCase
 
         $this->assertTrue($config['profiling']['enabled']);
         $this->assertTrue($config['fake_ip']['enabled']);
-        $this->assertEquals('33.33.33.11', $config['fake_ip']['ip']);
+        $this->assertSame('33.33.33.11', $config['fake_ip']['ip']);
     }
 
     public function testGetConfigTreeBuilderNoDebug()
