+++ composer.json
@@ -14,9 +14,9 @@
     "php": "^7.1",
     "geocoder-php/plugin": "^1.0",
     "php-http/discovery": "^1.7",
-    "symfony/console": "^3.4 || ^4.2",
-    "symfony/framework-bundle": "^3.4 || ^4.2",
-    "symfony/options-resolver": "^3.4 || ^4.2",
+    "symfony/console": "^3.4 || ^4.2 || ^5.0",
+    "symfony/framework-bundle": "^3.4 || ^4.2 || ^5.0",
+    "symfony/options-resolver": "^3.4 || ^4.2 || ^5.0",
     "willdurand/geocoder": "^4.0"
   },
   "require-dev": {
@@ -52,12 +52,12 @@
     "geoip/geoip": "~1.17",
     "nyholm/nsa": "^1.1",
     "nyholm/psr7": "^1.0",
-    "nyholm/symfony-bundle-test": "^1.3",
+    "nyholm/symfony-bundle-test": "^1.6.1",
     "php-http/curl-client": "^1.7",
     "php-http/message": "^1.5",
-    "symfony/phpunit-bridge": "^4.3",
-    "symfony/validator": "^3.4 || ^4.2",
-    "symfony/yaml": "^3.4 || ^4.2"
+    "symfony/phpunit-bridge": "^4.4",
+    "symfony/validator": "^3.4 || ^4.2 || ^5.0",
+    "symfony/yaml": "^3.4 || ^4.2 || ^5.0"
   },
   "conflict":  {
     "willdurand/geocoder": "3.2",
