+++ phpstan.neon.dist
@@ -10,5 +10,8 @@ parameters:
 			path: %currentWorkingDirectory%/DependencyInjection/Configuration.php
 			message: '#Call to an undefined method Symfony\\Component\\Config\\Definition\\Builder\\NodeParentInterface::(end|variableNode)#'
 		-
+			path: %currentWorkingDirectory%/DependencyInjection/Configuration.php
+			message: '#Call to an undefined method Symfony\\Component\\Config\\Definition\\Builder\\TreeBuilder::root\(\)#'
+		-
 			path: %currentWorkingDirectory%/ProviderFactory/GeoipFactory.php
 			message: '#Geocoder\\Provider\\Geoip\\Geoip not found#'
